Virus now implements IVirus interface

+++ core/src/com/mygdx/chalmersdefense/model/IMapObject.java
@@ -12,5 +12,5 @@ public interface IMapObject {
 
     String getSpriteKey();
 
-    int getAngle();
+    float getAngle();
 }
